Update and rename codeacross2016 to 2016-02-23-codeacross2016.md
Changed the file name and extension for proper routing and page display.

Made a variety of copy edits, some to improve markdown display, some to match press release conventions, and some to correct typos or make opinionated changes to phrasing I thought sounded better.  Added a quoted paragraph from the CfA sample because quotes are always nice to have in press releases.

Much of this is open to discussion.