Add client-side redirect to opensmc.tech
Since we do not have access to opensmc.org's domain records, we've moved our site to opensmc.tech. DNS records for opensmc.org point to GitHub pages, which doesn't allow for server-side redirects, so we're forced to handle it client-side.

This uses `location.replace`, which redirects the browser to the new location, and avoids saving the old location in the browser history. This won't send a 301 redirect response, however, so I've added a canonical link tag for search engine indexing.